service-service/requirements.txt
# RedFreelance/service-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.30
python-dotenv==1.0.1